breakout individual tasks
