+++ src/main/scala/com/localytics/sbt/sqs/ElasticMQPlugin.scala
@@ -1,7 +1,7 @@
 package com.localytics.sbt.sqs
 
 import com.localytics.sbt.sqs.ElasticMQKeys._
-import com.localytics.sbt.sqs.ElasticMQTasks._
+import sbt.Keys._
 import sbt._
 
 object ElasticMQPlugin extends AutoPlugin {
@@ -25,9 +25,10 @@ object ElasticMQPlugin extends AutoPlugin {
     restSQSConf             := RestSQSConf(),
     queuesConf              := Seq(),
 
-    downloadElasticMQ       <<= downloadElasticMQTask,
-    startElasticMQ          <<= startElasticMQTask,
-    stopElasticMQ           <<= stopElasticMQTask,
-    elasticMQTestCleanup    <<= elasticMQTestCleanupTask
+    downloadElasticMQ       := DownloadElasticMQ(elasticMQVersion.value, elasticMQUrl.value, elasticMQDir.value, elasticMQFileName.value, streams.value),
+    startElasticMQ          := StartElasticMQ(elasticMQDir.value, elasticMQFileName.value, elasticMQHeapSize.value, nodeAddressConf.value, restSQSConf.value, queuesConf.value, streams.value),
+    stopElasticMQ           := StopElasticMQ(streams.value),
+    elasticMQTestCleanup    := Tests.Cleanup(() => StopElasticMQ(streams.value)),
+    startElasticMQ          := startElasticMQ.dependsOn(downloadElasticMQ).value
   )
 }
